# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API bitkakir project",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/grid-settings": {
            "get": {
                "tags": [
                    "Grid settings"
                ],
                "summary": "List grid settings",
                "description": "Display listing of the resource",
                "operationId": "c3f9df71df1b70b6e52661f26c3c530f",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Grid settings"
                ],
                "summary": "Save account grid settings",
                "description": "Creates and returns a new grid settings",
                "operationId": "f16f3b9f7183cd8fa5cc851f84783a32",
                "requestBody": {
                    "$ref": "#/components/requestBodies/GridSettingsRequest"
                },
                "responses": {
                    "201": {
                        "description": "Category created successfully"
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                }
            }
        },
...
        "/api/v1/favorites/search/{id}": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Show favorite by ID of user",
                "description": "Display a listing by ID of user",
                "operationId": "d4ff3d02e28878e31acde4696d9b5c72",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search by ID of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/favorites/store": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Save favorite",
                "description": "Creates and returns a new favorite",
                "operationId": "6944cf00a6dfd4fb0cb9c78b0a0bd74d",
                "requestBody": {
                    "description": "Relationship between user, product, and service",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "User ID (must exist in the users table)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "product_id": {
                                        "description": "Product ID (optional, must exist in the products table)",
                                        "type": "integer",
                                        "example": 10,
                                        "nullable": true
                                    },
                                    "service_id": {
                                        "description": "Service ID (optional, must exist in the services table)",
                                        "type": "integer",
                                        "example": 5,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Favorite created successfully"
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                }
            }
        },
        "/api/v1/favorites/delete/{id}": {
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Remove the specified resource from storage",
                "description": "Delete favorite",
                "operationId": "fccf4f74c1d63b1a45ba0c9a47b022b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Delete by ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/api/v1/grid-settings/{id}": {
            "get": {
                "tags": [
                    "Grid settings"
                ],
                "summary": "Show account grid settings by ID",
                "description": "Display a listing by ID",
                "operationId": "bed8db8878016ee3aafe5de0dbf0d887",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search by ID",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Grid settings"
                ],
                "summary": "Update account grid settings",
                "description": "Update the specified resource in storage",
                "operationId": "820d0c8b2e9fdfdca19d5112039aca94",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search by ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/GridSettingsRequest"
                },
                "responses": {
                    "201": {
                        "description": "Category updated successfully"
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Grid settings"
                ],
                "summary": "Remove the specified resource from storage",
                "description": "Delete account grid settings",
                "operationId": "7166102d657f22c15686462df6804b38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Delete by ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "requestBodies": {
...
            "GridSettingsRequest": {
                "description": "Grid configuration parameters",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "cross_axis_count": {
                                    "description": "Number of items along the cross axis",
                                    "type": "integer",
                                    "example": 3
                                },
                                "child_aspect_ratio": {
                                    "description": "Aspect ratio of each child item",
                                    "type": "number",
                                    "format": "float",
                                    "example": 1.5
                                },
                                "main_axis_spacing": {
                                    "description": "Spacing along the main axis",
                                    "type": "integer",
                                    "example": 10
                                },
                                "cross_axis_spacing": {
                                    "description": "Spacing along the cross axis",
                                    "type": "integer",
                                    "example": 8
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        }
    },
...
}